INGM-391 Move PDO code to its own module
