+++ docs/ingeniamotion/capture/pdo.rst
@@ -1,0 +1,5 @@
+PDO
+===
+
+.. automodule:: ingeniamotion.pdo
+   :members:
